Update the 404 page.
Note: this goes along with a script on the webserver to convert all
the links/refs to absolute paths (e.g. starts with '/') so that
when used as an ErrorDocument (as it is), and it can be accessed
from multiple paths (as might happen) the page still works right
if it is not a top level 404.

This keeps the www.geomoose.org implementation detail of where the
site is located and how ErrorDocument is used as an implementation
detail for www.geomoose.org and not leaking all over the place.

This also keeps the normal pages all relative so they are easy to
test/stage at paths other than the top of a web server.

And it lets us build the 404.html page with the rest of the site
so that the styling and links stay consistent and up to date with
the rest of the site.